+++ src/driver/sqljs/SqljsDriver.ts
@@ -248,7 +248,7 @@ export class SqljsDriver extends AbstractSqliteDriver {
     protected async createDatabaseConnectionWithImport(database?: Uint8Array): Promise<any> {
         // sql.js < 1.0 exposes an object with a `Database` method.
         const isLegacyVersion = typeof this.sqlite.Database === "function";
-        const sqlite = isLegacyVersion ? this.sqlite : await this.sqlite(); 
+        const sqlite = isLegacyVersion ? this.sqlite : await this.sqlite(this.options.sqlJsConfig);
         if (database && database.length > 0) {
             this.databaseConnection = new sqlite.Database(database);
         }